Move MySQL to autoloading class
