+++ www/includes/easyparliament/init.php
@@ -79,9 +79,7 @@ if (!isset($_SERVER['WINDIR'])) {
 
 $DATA = new \MySociety\TheyWorkForYou\Data;
 
-include_once (INCLUDESPATH."mysql.php");
-
-Class ParlDB extends MySQL {
+class ParlDB extends \MySociety\TheyWorkForYou\Db\Connection {
     public function __construct() {
         $this->init (OPTION_TWFY_DB_HOST, OPTION_TWFY_DB_USER, OPTION_TWFY_DB_PASS, OPTION_TWFY_DB_NAME);
     }
